refactor: replace legacy ErrorMessage.js with messages/ErrorMessage.tsx
