+++ frontend/web/components/modals/CreateEditIntegrationModal.tsx
@@ -2,7 +2,7 @@ import React, { FC, FormEvent, useEffect, useState } from 'react'
 import EnvironmentSelect from 'components/EnvironmentSelect'
 import MyGitHubRepositoriesComponent from 'components/MyGitHubRepositoriesComponent'
 import _data from 'common/data/base/_data'
-import ErrorMessage from 'components/ErrorMessage'
+import ErrorMessage from 'components/messages/ErrorMessage'
 import Button from 'components/base/forms/Button'
 import classNames from 'classnames'
 import { getStore } from 'common/store'
