+++ frontend/web/components/modals/CreateUser.tsx
@@ -1,6 +1,6 @@
 import React, { FC, useEffect, useState } from 'react'
 import ChipInput from 'components/ChipInput'
-import ErrorMessage from 'components/ErrorMessage'
+import ErrorMessage from 'components/messages/ErrorMessage'
 import Button from 'components/base/forms/Button'
 import { useCreateIdentitiesMutation } from 'common/services/useIdentity'
 import Utils from 'common/utils/utils'
