+++ frontend/web/components/modals/CreateWebhook.tsx
@@ -2,7 +2,7 @@ import Constants from 'common/constants'
 import { useGetEnvironmentQuery } from 'common/services/useEnvironment'
 import { Webhook } from 'common/types/responses'
 import Utils from 'common/utils/utils'
-import ErrorMessage from 'components/ErrorMessage'
+import ErrorMessage from 'components/messages/ErrorMessage'
 import Highlight from 'components/Highlight'
 import TestWebhook from 'components/TestWebhook'
 import ViewDocs from 'components/ViewDocs'
