+++ frontend/web/components/modals/InviteUsers.tsx
@@ -10,7 +10,7 @@ import InputGroup from 'components/base/forms/InputGroup'
 import OrganisationProvider from 'common/providers/OrganisationProvider'
 import Utils from 'common/utils/utils'
 import _ from 'lodash'
-import ErrorMessage from 'components/ErrorMessage'
+import ErrorMessage from 'components/messages/ErrorMessage'
 import AccountStore from 'common/stores/account-store'
 import { close as closeIcon } from 'ionicons/icons'
 import { MultiValueProps } from 'react-select'
