+++ frontend/web/components/modals/create-feature/tabs/CreateFeatureTab.tsx
@@ -2,7 +2,7 @@ import React, { FC, useCallback, useEffect, useState } from 'react'
 import { FeatureState, ProjectFlag } from 'common/types/responses'
 import FeatureValueTab from './FeatureValueTab'
 import FeatureSettingsTab from './FeatureSettingsTab'
-import ErrorMessage from 'components/ErrorMessage'
+import ErrorMessage from 'components/messages/ErrorMessage'
 import WarningMessage from 'components/WarningMessage'
 import { useHasPermission } from 'common/providers/Permission'
 import { ProjectPermission } from 'common/types/permissions.types'
