+++ frontend/web/components/pages/AccountSettingsPage.js
@@ -1,7 +1,7 @@
 // import propTypes from 'prop-types';
 import React, { Component, ReactNode } from 'react'
 import Button from 'components/base/forms/Button'
-import ErrorMessage from 'components/ErrorMessage'
+import ErrorMessage from 'components/messages/ErrorMessage'
 import _data from 'common/data/base/_data'
 import ConfigProvider from 'common/providers/ConfigProvider'
 import TwoFactor from 'components/TwoFactor'
