+++ frontend/web/components/pages/ChangeRequestDetailPage.tsx
@@ -40,7 +40,7 @@ import MyGroupsSelect from 'components/MyGroupsSelect'
 import Panel from 'components/base/grid/Panel'
 import DiffChangeRequest from 'components/diff/DiffChangeRequest'
 import JSONReference from 'components/JSONReference'
-import ErrorMessage from 'components/ErrorMessage'
+import ErrorMessage from 'components/messages/ErrorMessage'
 import ConfigProvider from 'common/providers/ConfigProvider'
 import { useHistory } from 'react-router-dom'
 import { openPublishChangeRequestConfirm } from 'components/PublishChangeRequestModal'
