+++ frontend/web/components/pages/FeatureHistoryDetailPage.tsx
@@ -12,7 +12,7 @@ import { Environment } from 'common/types/responses'
 import PageTitle from 'components/PageTitle'
 import FeatureVersion from 'components/FeatureVersion'
 import moment from 'moment'
-import ErrorMessage from 'components/ErrorMessage'
+import ErrorMessage from 'components/messages/ErrorMessage'
 import Tabs from 'components/navigation/TabMenu/Tabs'
 import TabItem from 'components/navigation/TabMenu/TabItem'
 import Breadcrumb from 'components/Breadcrumb'
