+++ frontend/web/components/pages/UsersAndPermissionsPage.tsx
@@ -26,7 +26,7 @@ import { useGetRolesQuery } from 'common/services/useRole'
 import AppActions from 'common/dispatcher/app-actions'
 import { RouterChildContext, useHistory } from 'react-router-dom'
 import Input from 'components/base/forms/Input'
-import ErrorMessage from 'components/ErrorMessage'
+import ErrorMessage from 'components/messages/ErrorMessage'
 import PanelSearch from 'components/PanelSearch'
 import moment from 'moment'
 import sortBy from 'lodash/sortBy'
