+++ frontend/web/components/tags/CreateEditTag.tsx
@@ -13,7 +13,7 @@ import InputGroup from 'components/base/forms/InputGroup'
 import Button from 'components/base/forms/Button'
 import Tag from './Tag'
 import InlineModal from 'components/InlineModal'
-import ErrorMessage from 'components/ErrorMessage'
+import ErrorMessage from 'components/messages/ErrorMessage'
 import Switch from 'components/Switch'
 import Icon from 'components/Icon'
 import { ProjectPermission } from 'common/types/permissions.types'
